brainmap-deploy/requirements.txt
fastapi==0.115.6
fastjsonschema==2.21.1
uvicorn==0.34.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.8.2
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.30.1
google-cloud-core==2.4.1
google-cloud-storage==2.5.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.56.4
matplotlib==3.9.3
matplotlib-inline==0.1.7
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy==1.23.4
pandas==2.2.3
pillow==11.0.0
protobuf==3.20.3
pytest==8.3.4
shap==0.46.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.16.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
ultralytics==8.3.49
ultralytics-thop==2.0.13
colorama==0.4.6
prefect
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
streamlit==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.19
